requirements.txt
numpy==1.24.4
pandas==2.0.3
python-dotenv==1.0.1
matplotlib==3.9.0

# sudo apt install xterm
# sudo apt install dosunix
# sudo apt-get install tmux